0.99
Structure and routes changes
Template cleanup
Admin messages
Dashboard functionality restored

+++ app/Http/Controllers/UserTypeController.php
@@ -55,8 +55,10 @@ class UserTypeController extends Controller
 
     public function deleteSectors(Request $request){
         foreach($request->ids as $id){
+          if($id != 1){
             $type = UserType::find($id);
             $type->delete();
+          }
         }
 
         return back();
